refactor: 디렉토리 이동 및 상수 네이밍 변경
