+++ Assignment/blog/src/main/java/com/sopt/org/common/dto/SuccessStatusResponse.java
@@ -1,5 +1,7 @@
 package com.sopt.org.common.dto;
 
+import com.sopt.org.common.dto.message.SuccessMessage;
+
 public record SuccessStatusResponse(
         int status,
         String message
@@ -8,5 +10,4 @@ public record SuccessStatusResponse(
     public static SuccessStatusResponse of(SuccessMessage successMessage) {
         return new SuccessStatusResponse(successMessage.getStatus(), successMessage.getMessage());
     }
-
 }
