+++ Assignment/blog/src/main/java/com/sopt/org/common/dto/message/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.sopt.org.exception.message;
+package com.sopt.org.common.dto.message;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import org.springframework.http.HttpStatus;
 @Getter
 public enum ErrorMessage {
     MEMBER_NOT_FOUND_BY_ID_EXCEPTION(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 사용자가 존재하지 않습니다."),
-    BLOG_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "블로그가 존재하지 않습니다"),
+    BLOG_NOT_FOUND_BY_ID_EXCEPTION(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 블로그가 존재하지 않습니다"),
     POST_NOT_FOUND_BY_ID_EXCEPTION(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 포스트가 존재하지 않습니다");
 
     private final int status;
