+++ Assignment/blog/src/main/java/com/sopt/org/exception/BusinessException.java
@@ -1,6 +1,6 @@
 package com.sopt.org.exception;
 
-import com.sopt.org.exception.message.ErrorMessage;
+import com.sopt.org.common.dto.message.ErrorMessage;
 import lombok.Getter;
 
 @Getter
