+++ Assignment/blog/src/main/java/com/sopt/org/exception/NotFoundException.java
@@ -1,6 +1,6 @@
 package com.sopt.org.exception;
 
-import com.sopt.org.exception.message.ErrorMessage;
+import com.sopt.org.common.dto.message.ErrorMessage;
 
 public class NotFoundException extends BusinessException {
     public NotFoundException(ErrorMessage errorMessage) {
